Rewrite most tests in OpenSPF format
