+++ tests/rfc4408/rfc4408-ignore.yml
@@ -1,0 +1,1 @@
+- 'exp-only-macro-char'
