+++ tests/rfc7208/rfc7208-ignore.yml
@@ -1,0 +1,2 @@
+- 'exp-only-macro-char'
+- 'spftimeout'
